+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -51,6 +51,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         logger: logging.Logger,
         command_tracker,
         connection_state_callback,
+        quality_state_callback,
         tango_device_name: str,
         ds_device_fqdn: str,
         spf_device_fqdn: str,
@@ -89,7 +90,10 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             spfconnectionstate=CommunicationStatus.NOT_ESTABLISHED,
             spfrxconnectionstate=CommunicationStatus.NOT_ESTABLISHED,
             dsconnectionstate=CommunicationStatus.NOT_ESTABLISHED,
+            band1pointingmodelparams=[],
             band2pointingmodelparams=[],
+            band3pointingmodelparams=[],
+            band4pointingmodelparams=[],
             trackinterpolationmode=None,
             ignorespf=None,
             ignorespfrx=None,
@@ -97,6 +101,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         )
         self.logger = logger
         self._connection_state_callback = connection_state_callback
+        self._quality_state_callback = quality_state_callback
         self._dish_mode_model = DishModeModel()
         self._state_transition = StateTransition()
         self._command_tracker = command_tracker
@@ -123,6 +128,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     self._sub_communication_state_changed, "spfConnectionState"
                 ),
                 component_state_callback=self._component_state_changed,
+                quality_state_callback=self._quality_state_callback,
             ),
             "DS": DSComponentManager(
                 ds_device_fqdn,
@@ -139,12 +145,16 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 achievedpointing=[0.0, 0.0, 0.0],
                 achievedpointingaz=[0.0, 0.0, 0.0],
                 achievedpointingel=[0.0, 0.0, 0.0],
+                band1pointingmodelparams=[],
                 band2pointingmodelparams=[],
+                band3pointingmodelparams=[],
+                band4pointingmodelparams=[],
                 trackinterpolationmode=TrackInterpolationMode.SPLINE,
                 communication_state_callback=partial(
                     self._sub_communication_state_changed, "dsConnectionState"
                 ),
                 component_state_callback=self._component_state_changed,
+                quality_state_callback=self._quality_state_callback,
             ),
             "SPFRX": SPFRxComponentManager(
                 spfrx_device_fqdn,
@@ -167,6 +177,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                     self._sub_communication_state_changed, "spfrxConnectionState"
                 ),
                 component_state_callback=self._component_state_changed,
+                quality_state_callback=self._quality_state_callback,
             ),
         }
         initial_component_states = {
@@ -184,7 +195,10 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             "spfconnectionstate": CommunicationStatus.NOT_ESTABLISHED,
             "spfrxconnectionstate": CommunicationStatus.NOT_ESTABLISHED,
             "dsconnectionstate": CommunicationStatus.NOT_ESTABLISHED,
+            "band1pointingmodelparams": [],
             "band2pointingmodelparams": [],
+            "band3pointingmodelparams": [],
+            "band4pointingmodelparams": [],
             "ignorespf": False,
             "ignorespfrx": False,
         }
@@ -386,16 +400,6 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 ds_component_state["pointingstate"],
             )
             self._update_component_state(pointingstate=ds_component_state["pointingstate"])
-
-            if ds_component_state["pointingstate"] in [
-                PointingState.SLEW,
-                PointingState.READY,
-            ]:
-                # TODO ST (04/2024) achievedtargetlock needs to be determined
-                # from configured threshold, see configureTargetLock
-                self._update_component_state(achievedtargetlock=False)
-            elif ds_component_state["pointingstate"] == PointingState.TRACK:
-                self._update_component_state(achievedtargetlock=True)
 
         # spf bandInFocus
         if not self.is_device_ignored("SPF") and (
@@ -610,6 +614,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             self._dish_mode_model.is_command_allowed,
             "SetStandbyLPMode",
             component_manager=self,
+            task_callback=task_callback,
         )
 
         status, response = self.submit_task(
@@ -629,6 +634,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             self._dish_mode_model.is_command_allowed,
             "SetStandbyFPMode",
             component_manager=self,
+            task_callback=task_callback,
         )
 
         status, response = self.submit_task(
@@ -649,6 +655,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             self._dish_mode_model.is_command_allowed,
             "SetOperateMode",
             component_manager=self,
+            task_callback=task_callback,
         )
         status, response = self.submit_task(
             self._command_map.set_operate_mode,
@@ -711,6 +718,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             self._dish_mode_model.is_command_allowed,
             req_cmd,
             component_manager=self,
+            task_callback=task_callback,
         )
 
         status, response = self.submit_task(
@@ -730,6 +738,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             self._dish_mode_model.is_command_allowed,
             "SetStowMode",
             component_manager=self,
+            task_callback=task_callback,
         )
         status, response = self.submit_task(
             self._command_map.set_stow_mode,
@@ -811,12 +820,21 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         self,
         k_value,
     ) -> Tuple[ResultCode, str]:
-        """Set the k-value on the SPFRx"""
+        """Set the k-value on the SPFRx.
+        Note that it will only take effect after
+        SPFRx has been restarted.
+        """
         spfrx_cm = self.sub_component_managers["SPFRX"]
         try:
-            spfrx_cm.write_attribute_value("kvalue", k_value)
-        except LostConnection:
-            return (ResultCode.REJECTED, "Lost connection to SPFRx")
+            result = spfrx_cm.execute_command("SetKValue", k_value)
+            self.logger.debug(
+                "Result of the call to [%s] on SPFRx is [%s]",
+                "SetKValue",
+                result,
+            )
+        except (LostConnection, tango.DevFailed) as err:
+            self.logger.exception("SetKvalue on SPFRx failed")
+            return (ResultCode.FAILED, err)
         return (ResultCode.OK, "Successfully requested SetKValue on SPFRx")
 
     def set_track_interpolation_mode(