+++ src/ska_mid_dish_manager/component_managers/spf_cm.py
@@ -1,4 +1,5 @@
 """Specialization for SPF functionality"""
+
 import logging
 from threading import Lock
 from typing import Any, Callable, Optional