+++ tests/acceptance/conftest.py
@@ -9,7 +9,6 @@ from ska_mid_dish_manager.models.dish_enums import (
     IndexerPosition,
     SPFOperatingMode,
 )
-from tests.utils import set_ignored_devices
 
 
 @pytest.fixture
@@ -40,10 +39,6 @@ def setup_and_teardown(
     )
     event_store.wait_for_value((), timeout=30)
     event_store.clear_queue()
-
-    set_ignored_devices(
-        dish_manager_proxy=dish_manager_proxy, ignore_spf=False, ignore_spfrx=False
-    )
 
     spfrx_device_proxy.ResetToDefault()
     spf_device_proxy.ResetToDefault()