+++ tests/acceptance/test_scan_and_end_scan_commands.py
@@ -1,4 +1,5 @@
 """Test the Scan and EndScan command"""
+
 import pytest
 import tango