+++ tests/acceptance/test_stow.py
@@ -1,4 +1,5 @@
 """Test that DS goes into STOW and dishManager reports it"""
+
 import pytest
 import tango